+++ implementations/rust/ockam/ockam_channel/src/error.rs
@@ -1,9 +1,6 @@
 use ockam_core::{
-    error::{
-        code::{ErrorCode, Kind, Origin},
-        Error2,
-    },
-    thiserror,
+    errcode::{ErrorCode, Kind, Origin},
+    thiserror, Error2,
 };
 
 /// Types of errors that may occur constructing a secure channel.
