+++ implementations/rust/ockam/ockam_core/src/routing/error.rs
@@ -1,5 +1,5 @@
-use crate::error::{
-    code::{ErrorCode, Kind, Origin},
+use crate::{
+    errcode::{ErrorCode, Kind, Origin},
     Error2,
 };
