+++ implementations/rust/ockam/ockam_identity/src/change.rs
@@ -1,6 +1,5 @@
 use ockam_core::compat::vec::Vec;
-use ockam_core::vault::PublicKey;
-use ockam_core::Result;
+use ockam_core::{vault::PublicKey, Result};
 use serde::{Deserialize, Serialize};
 
 pub use crate::signature::*;
