+++ implementations/rust/ockam/ockam_key_exchange_xx/src/responder.rs
@@ -4,8 +4,7 @@ use ockam_core::compat::{
     string::{String, ToString},
     vec::Vec,
 };
-use ockam_core::error::Result;
-use ockam_core::{async_trait, compat::boxed::Box};
+use ockam_core::{Result, async_trait, compat::boxed::Box};
 use ockam_key_exchange_core::{CompletedKeyExchange, KeyExchanger};
 
 #[derive(Debug)]
