+++ implementations/rust/ockam/ockam_node/src/delayed.rs
@@ -1,7 +1,7 @@
 use crate::Context;
 use core::time::Duration;
 use futures::future::{AbortHandle, Abortable};
-use ockam_core::{error::Result, Address, Message};
+use ockam_core::{Result, Address, Message};
 
 /// Allow to send message to destination address periodically after some delay
 /// Only one scheduled heartbeat allowed at a time
