+++ implementations/rust/ockam/ockam_node/src/router/start_processor.rs
@@ -1,7 +1,7 @@
 use super::{AddressMeta, AddressRecord, NodeState, Router, SenderPair};
 use crate::tokio::sync::mpsc::Sender;
 use crate::{error, NodeReply, NodeReplyResult, Reason};
-use ockam_core::{error::Result, Address};
+use ockam_core::{Result, Address};
 
 /// Execute a `StartWorker` command
 pub(super) async fn exec(
