+++ implementations/rust/ockam/ockam_node/src/router/start_worker.rs
@@ -1,7 +1,7 @@
 use super::{AddressMeta, AddressRecord, NodeState, Router, SenderPair};
 use crate::tokio::sync::mpsc::Sender;
 use crate::{error, NodeReply, NodeReplyResult, Reason};
-use ockam_core::{error::Result, AddressSet};
+use ockam_core::{Result, AddressSet};
 
 /// Execute a `StartWorker` command
 pub(super) async fn exec(
