+++ implementations/rust/ockam/ockam_node/src/router/utils.rs
@@ -1,7 +1,7 @@
 use super::Router;
 use crate::tokio::sync::mpsc::Sender;
 use crate::{error, NodeReply, NodeReplyResult, Reason};
-use ockam_core::{error::Result, Address, TransportType};
+use ockam_core::{Address, Result, TransportType};
 
 /// Receive an address and resolve it to a sender
 ///
