+++ implementations/rust/ockam/ockam_vault/src/signer_impl.rs
@@ -1,8 +1,7 @@
 use crate::vault::Vault;
 use crate::VaultError;
 use ockam_core::vault::{Secret, SecretType, Signature, Signer};
-use ockam_core::error::Result;
-use ockam_core::{async_trait, compat::boxed::Box};
+use ockam_core::{async_trait, compat::boxed::Box, Result};
 
 #[async_trait]
 impl Signer for Vault {
